scripts/requirements.txt
docopt==0.6.2
Jinja2
python-dotenv
